changed flag for mount and moved files
