+++ cloudos_cli/_version.py
@@ -1,1 +1,1 @@
-__version__ = '2.91.0'
+__version__ = '2.92.0'
