+++ cloudos_cli/interactive_session/__init__.py
@@ -1,1 +1,5 @@
 """CloudOS interactive session module."""
+
+from .link import Link
+
+__all__ = ['Link']
