--- cloudos_cli/link/__init__.py
@@ -1,8 +1,0 @@
-"""
-Functions and classes related to datasets.
-"""
-
-from .link import Link
-
-
-__all__ = ['link']
